requirements.txt
tabulate==0.9.0
requests==2.31.0
python-dotenv==1.0.0
deepdiff==6.7.1 